Create Palindrome_Program.cpp
This code first gets the word from the user. Then, it creates a boolean variable called isPalindrome and initializes it to true. Next, it iterates through the string, comparing each character to the character at the opposite end of the string. If any of the characters are not equal, it sets isPalindrome to false and breaks out of the loop. Finally, it prints a message indicating whether or not the word is a palindrome.